étape 10 corrections

+++ src/ch/epfl/rigel/gui/ViewingParametersBean.java
@@ -1,24 +1,28 @@
 package ch.epfl.rigel.gui;
 
 import ch.epfl.rigel.coordinates.HorizontalCoordinates;
-import javafx.beans.property.*;
+import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleObjectProperty;
 
+import java.time.LocalDate;
 
 
 public final class ViewingParametersBean {
 
-    private final IntegerProperty fieldOfViewDeg = new SimpleIntegerProperty();
+    private final DoubleProperty fieldOfViewDeg = new SimpleDoubleProperty();
     private final ObjectProperty<HorizontalCoordinates> center = new SimpleObjectProperty<>(null);
 
-    public IntegerProperty fieldOfViewDegProperty(){
+    public DoubleProperty fieldOfViewDegProperty(){
         return fieldOfViewDeg;
     }
 
-    public int getFieldOfViewDeg(){
+    public Double getFieldOfViewDeg(){
         return fieldOfViewDeg.get();
     }
 
-    public void setFieldOfViewDeg(int fieldOfViewDeg){
+    public void setFieldOfViewDeg(Double fieldOfViewDeg){
         this.fieldOfViewDeg.set(fieldOfViewDeg);
     }
